requirements.txt
streamlit
groq
langchain
langgraph
langchain-groq
langchain-community
sentence-transformers
faiss-cpu
pandas
numpy
python-dotenv
mlflow
requests
